requirements.txt
alembic==1.12.0
annotated-types==0.6.0
anyio==3.7.1
billiard==4.1.0
celery==5.3.4
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
exceptiongroup==1.1.3
fastapi==0.104.0
greenlet==3.0.0
h11==0.14.0
httptools==0.6.1
idna==3.4
kombu==5.3.2
Mako==1.2.4
MarkupSafe==2.1.3
migrate==0.3.8
paho-mqtt==1.6.1
prompt-toolkit==3.0.39
psycopg2-binary==2.9.9
pydantic==2.4.2
pydantic_core==2.10.1
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0.1
schedule==1.2.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.22
starlette==0.27.0
typing_extensions==4.8.0
tzdata==2023.3
uvicorn==0.23.2
uvloop==0.18.0
vine==5.0.0
watchfiles==0.21.0
wcwidth==0.2.8
websockets==11.0.3